requirements.txt
streamlit
pandas
google.generativeai
io
contextlib